requirements.txt
python-dotenv
requests
streamlit
colour